Soul EV - 64kWh: Add temperature of module 5 (experimental)
I observed that the maximal battery temperature was below the maximum of
the modules 1-4. Also, it's known that the 64 kW model has _five_
modules.

I concluded that somewhere there might be the temperature of the fifth
module. By looking at a couple scans which I found online, I located a
candidate which I added with a `00X` prefix (`X` stands for
eXperimental).